2022-01-22
Saturday, January 22nd, 2022 - the Meat Loaf show! With the unpleasant and yet inevitable arrival of another notification of an artist passing, came many messages from friends and family asking about a Meat Loaf show, which demonstrates how far reaching his appeal was. A guilty pleasure for some, and a top ten artist for others, Meat Loaf was certainly musical Marmite. I love the theatricality and the pure performance, but I didn't know too much outside of the Bat Out of Hell record - which is one of my absolute favourites! Today's show is a play-through of Meat Loaf from his earliest recordings to his last album. Here's to hoping that he and Jim Steinman are working on new material together again.