2022-02-11
Friday, February 11th, 2021 - the better late than never new music show! I've had this playlist ready for a month or so now, so it's not as brand new as previous new music shows, but nonetheless it does contain tracks that I've been looking forward to playing out. Upbeat and punky in places, and mellow and introspective in others, if you find yourself wondering where to find new music, it's here on a Friday this time!