2022-02-17
Thursday, February 17th, 2021 - the new-ish music show! Selections from a "to play" playlist tonight. By no design, a lot of tonight's tracks come from around 2019, and most if not all could be considered new-wave of post-punk, upbeat, urgent, tongue-in-cheek, fun, funky and fresh. Not heavy by any means, but the tracks on the more gentle side are a little more synthy, or truer to the classic 90s US alternative sound. Stick it on, turn it up, and hopefully discover something new!