2022-02-18
Friday, February 18th, 2021 - the Celtic punk, folk rock show! Returning to the traditional way of playlisting a show based on the first song in my head when I wake up, today was Dropkick Murphy's I'm Shipping Up To Boston, that was used brilliantly in the 2006 film The Departed. I knew they had their contemporaries, but wondered if there was enough of them to fill an entire show. Today is an exploration of that wonder. Not designed as a drinking show, but inevitably turning out to be, it also fits the "Friday your Friday" mentally that is so often necessary. Cheers!