2022-02-22
Tuesday, February 22nd, 2022 - the rarities, re-issues, and new releases show! All "new" music released in the last month or so whether that be a demo from a deluxe version of an album, a re-mastered classic, or brand new music from the last couple of weeks. Anyone notice anything interesting about the date today? This show did air as the clocks read 22:22 22/2/22 in the UK. Was there a magical magnetic event, or was that just Frank Zappa and the Mother's of Invention's Magic Fingers getting to you?