2022-02-23
Wednesday, February 23rd, 2021 - the Summer of Love show! Today's show was inspired by a tweet shared with me this morning by my Dad containing lyrics to Procol Harum's A Whiter Shade of Pale, containing a rarely heard third verse. As this track was one of the anthems of the Summer of Love, I decided to base the show around it, working chronologically from January to August thinking about what it must have been like to have been buying albums in the run up to the Summer. I've tried to play deeper cuts from the albums as I like to imagine them ringing distant bells for people who owned them at the time, instead of playing through the more well known singles. With the passing of Gary Brooker on Saturday, A Whiter Shade of Pale deserves the extra attention and after a few hours of back and forth, my Dad shared a version of the song with a fourth rarely heard verse recorded live for its 30th anniversary in 1997.