2022-02-25
Friday, February 25th, 2022 - the Saharan Desert Blues show! In a departure from the regular "It's Friday, let's party" vibe usually associated with the show, it felt more appropriate to use music as a distraction tonight. I found an intriguing playlist on Tidal called Desert Electric: Saharan Guitars and that was enough to get me interested and feeling challenged. From Tidal: "The culture of Sahara, from the Atlantic Ocean to the Red Sea, is as vast and wild as the desert itself. What might look like a deserted place on first sight, unveils a fertile soil at closer inspection. One of the most prominent styles of Saharan music – especially the Western parts including Mali and Niger and ethnic groups like Tuareg and Berber – is what is known as the Desert Blues."