2022-02-26
Saturday, February 26th, 2022 - the roots reggae show! Today's show is a simple, straightforward, 1970s reggae show. It was hard to pick a theme today, as world events are still a little overwhelming. It was however a beautiful day here and I found myself with a Bob Marley track to begin. From there it was a case of keeping up with the vibe. Deep, thundering bass-hug bass lines, snappy crackling snare drums, and enough reverb to float away in. Stick it on, turn it up, and let the pressure drop!