2022-03-01
Tuesday, March 1st, 2022 - the Seattle, Sub Pop, and the start of grunge show! Given the general mood of the world at the moment, I'm still finding myself looking for musical distractions, and after watching an episode of Dark Side of the 90's called Grunge & The Seattle Sound I was inspired to dip back into the genre that was always a bit of a musical side piece through my formative years. The early part of the show spends a little time exploring the late 80s and the origination of grunge, before settling in for some sonically overwhelming sludge to sink into.