2022-03-03
Thursday, March 3rd, 2022 - the early 00s alt rock, pop, and a little emo show! I was in a Jack Johnson mood this morning, but it's not always possible to stretch a vibe across a two hour show so I had to widen the criteria a little. Given as I played new music from Dashboard Confessional yesterday, if I went back to their early albums, that aligned perfectly with Jack, and placed me in the early 00s. There was a lot of great music for me at that time, and it was a great time in my life too, so as I continue to use music as an escape, it felt like the right year to travel back to tonight.