2022-03-06
Sunday, March 6th, 2022 - the blues rock, groove rock, Sunday swagger show! I like a lot of music, across a lot of genres, and a lot of decades, and I like discovering new music, and I like being introduced to new genres, but every now and again I need to return to my musical spirit animal. These are songs that have a swagger to them, they're confident and a little cheeky. This isn't a "my favourite songs" show, but it is a very happy vibe to settle into on a rainy Sunday. This one needs to be loud!