2022-03-12
Saturday, March 12th, 2022 - the hip hop, rap, r&b show! There's a very specific sort of sound I'm looking to listen to when I'm in the mood for this genre, and it isn't really summed up by the genre itself. Comedy hip hop might be the closest, but that makes it sound a bit too whackadoo. It's the sharp, quick-witted cheekiness of the lyrics, the irresistible funky nastiness of the track, and the overall attitude of it all wrapped (rapped?) up together. Essentially it seems to be anything that Dr Dre has been involved with, but of course there's more outside that. Today was inspired by a Snoop Dogg ear worm and the need to play more Lil' Kim. It's filthy, it's nasty, it's brilliant.