2022-03-13
Sunday, March 13th, 2022 - the soft rock, pop rock, not quite yacht rock show! Today's show is a big slick production, low BPM type of show that you can hopefully sit back and digest your roast dinner to. It pairs well with red wine and roast beef, but don't forget the horseradish because there always needs to be a little spice to a dinner (and a playlist). It's also a birthday extravaganza this week, so Happy Birthday again to all!