2022-03-19
Saturday, March 19th, 2022 - the yes it's more 90s alternative, but listen to the bass groove show! I'm a complete sucker for a big confident bass line, well placed in pristine 90s production. There's a comfort to it, and it's been described as a bass hug on the show before. This isn't necessarily prime bass hug territory, but if you do want to rattle some windows and move some air, you can crank this one and it shouldn't disappoint!