2022-07-22
Friday, July 22nd, 2022 - the heavy hitting 90s female vocalists towards the end of the decade show! To the best of my memory, there was a sweet spot in the early-mid 90s where female singer-songerwriters ruled the airwaves. From Shawn Colvin to Sheryl Crow, Fiona Apple to Alanis Morissette, and Tracy Bonham to Tori Amos - it was a great time for music. I've covered a fair amount of this material in shows focusing on Lilith Fair as well as the hits from the big albums, but what did their second, third, fourth and even tenth albums sound like? How did the early 90s sound evolve to close out the decade. Inspired by Fiona Apple's album When the Pawn... today's show is a play-through of exactly that.