2022-08-05
Friday, August 5th, 2022 - the 1988 hip hop hits show! The year was '88, Die Hard was the movie, table tennis the new Olympic sport, and CDs outsold vinyl for the first time. Today's show is a play-through of what was happening in hip hop the year that Hans Gruber uttered the immortal words: “Shoot the glass!”