2022-08-07
Sunday, August 7th, 2022 - the 18th birthday, songs from 2004 show! I think that 2004 was a great year for music. We were getting involved with a superb post-punk revival, and many of the more playful and edgier tracks came from a variety of genres making it into the mainstream. It was also the year I moved to America, and so I was experiencing and discovering music with a whole new group of people. But that's not the point of today's show. Today is also an 18th birthday show for Madeline. Here's what some of the world sounded like way back then with a selection of tracks from some of the biggest albums of the year.