2022-08-20
Saturday, August 20th, 2022 - the I was 18 in 1998 show! Last week we threw an epic 18th birthday/graduation party and it got me thinking about the sort of music that was everywhere when I was 18 in 1998. This is not a top ten, or number ones show today, instead an opportunity to drift back to the 90s in a musical time machine and sink into the sounds that may well have defined a year, a summer, or even just a monumentally epic weekend or night out. Time to dig out those low rise, boot cut jeans, sling on the ringer tee, frost your tips and pour yourself a WKD Blue to get the party started!