2022-08-21
Sunday, August 21st, 2022 - the rumblings of rock and roll show! I just watched the new Baz Luhrmann movie Elvis, which was excellent, and it inspired a desire to listen to some early Beale Street style blues. This show isn't an academic adventure, or an archaeological dig into the origins of rock and roll, rather it's an opportunity to listen to some of the music and artists that laid the foundations for what would become rock and roll, as well as some of the other developments within the genres of the time.