2022-09-18
Sunday, September 18th, 2022 - the mid-00s to mid-10s alt rock/power pop show! Sunny Sunday sounds from a decade of music that I was a little distant from at the time. In 2004 I moved from the UK to Memphis to become a Sound Designer at a live theatre, spending the next seven or so years fairly occupied and immersed in musical theatre. This is the "popular" indie music that mostly passed me by!