2022-11-17
Thursday, November 17th, 2022 - the two months old new music show! New music on a Thursday?! Today's show is a new music playlist originally scheduled for September that was never broadcast. It's still new, but it's losing a little of its new playlist smell. Heavy and rocky in places, light and poppy in others. Let's take it out for a ride, I promise it's still fresh!