2022-11-18
Friday, November 18th, 2022 - the new wave, synth pop, industrial, alternative show! It's finally Friday and this playlist has been brewing for a couple of weeks. I listened to Tears for Fears' Songs from the Big Chair loud on vinyl recently and it inspired a deep dive for some of the darker new wave, synth-pop, alternative sounds that were around in the mid-80s. Thumping bass lines, thundering drums, sinister synths, and oh-so 80s vocals. This is the mood!