2023-06-02
Friday, June 2nd, 2023 - the mid to late 70s ROCK show! Show number 420 was playlisted and ready for 4/20 (April 20th), but alas life intervened, and it has sat with me, teasing and tempting ever since. I have a huge soft spot for this kind of straightforward honest rock, and today's show is deeper cuts from well known artists, that most importantly had to groooooove. Summer's back, and to me this is perfect Friday night music to get your weekend going. See you at the bar, in the back yard, or by the pool!