2023-06-04
Sunday, June 4th, 2023 - the Guardian's of the Galaxy nine year old's birthday show! Our friend Wilf is nine years old today, and he loves Guardians of the Galaxy, Rollercoasters, and non-stop-dancing. Today's show takes a healthy dose of tracks from the three movies' mixtapes, and weaves in some Point Break picks to expand the vibe. This is a beer in the back garden BBQ playlist, so get your grills going and remember to stretch, because this one is a dancer! Happy Birthday Wilf!