2024-12-31
Tuesday, December 31st, 2024 - the New Year's Eve Rock & Roll Show! We're not getting into the birth of Rock and Roll here as a musical style, instead looking at its use in song titles and lyrics and how throughout the 60s, 70s, and 80s it became more of a mythology symbolising a lifestyle of rebellion, freedom, and energy! You know, the cool stuff! Happy New Year everyone!